Update poetry version
Poetry 1.1.0 changed how properties are ordered in the lock file. This change regenerates the lock file with the new dependency order but should not change anything.

poetry_core is now recommended for the build-system declaration. This makes tox and other tools faster as they don't need to download as much stuff.