Fix regex flags to fix warnings
Warnings like this show up when running tests in this project and in our internal projects that import this library:
    DeprecationWarning: Flags not at the start of the expression '\\b(?i)(\\d+)\\s{1,4}(['

This is because flags like `(?i)` which turns on case insensitive matching, actually apply to the entire regex and so are misleading if used in the middle. This has been true forever but the warning was just aded in Python 3.7. Moving them to the front of the string clarifies that they apply everywhere and removes the warning. This is not a functional change, the behavior is the same.

I ran some tests using hypothesis to verify that these regex are all equivalent to their previous versions. You can see the code I used to run that test on the `test_new_regexes` branch.

AIE-784